env系関数リファクタ
